requirements.txt
streamlit
pandas
numpy
pillow
scikit-learn
joblib
opencv-python-headless
scikit-image
scipy